refactor: 추상클래스를 사용해 딜러와 플레이어에서 발생하는 중복코드 제거
