+++ src/main/java/blackjack/domain/Dealer.java
@@ -5,25 +5,19 @@ import blackjack.domain.card.Deck;
 import blackjack.domain.card.Hand;
 import java.util.ArrayList;
 
-public class Dealer implements Gamer {
+public class Dealer extends Gamer {
 
     private static final int THRESHOLD = 16;
 
     private final Deck deck;
-    private final Hand hand;
 
     private Dealer(final Deck deck, final Hand hand) {
+        super(hand);
         this.deck = deck;
-        this.hand = hand;
     }
 
     public static Dealer from(final Deck deck) {
         return new Dealer(deck, new Hand(new ArrayList<>()));
-    }
-
-    public void drawInitialHand() {
-        draw();
-        draw();
     }
 
     public void draw() {
@@ -41,21 +35,6 @@ public class Dealer implements Gamer {
     @Override
     public boolean canDraw() {
         return hand.calculateOptimalSum() <= THRESHOLD;
-    }
-
-    @Override
-    public boolean isBust() {
-        return calculateScore() > BLACKJACK;
-    }
-
-    @Override
-    public boolean isBlackjack() {
-        return calculateScore() == BLACKJACK && hand.hasOnlyInitialCard();
-    }
-
-    @Override
-    public int calculateScore() {
-        return hand.calculateOptimalSum();
     }
 
     public Hand getHand() {
